Add conftest.py to avoid duplicate fixtures
